Move puzzle-sort-and-group to lib/
